+++ apps/demo/src/components/mm/lists/borrow/BorrowModal.tsx
@@ -189,7 +189,7 @@ export function BorrowModal({
           Number.parseFloat(
             formatUnits(BigInt(destinationMetrics.formattedReserve.availableLiquidity), aToken.decimals),
           ),
-          Number.parseInt(destinationMetrics.formattedReserve.borrowCap) -
+          Number.parseFloat(destinationMetrics.formattedReserve.borrowCap) -
             Number.parseFloat(destinationMetrics.formattedReserve.totalVariableDebt),
         ), 6);
       }
